1.完成了搜索结果页的数据匹配与加载更多. 2.解决一些导致App崩溃的BUG. 3.正式接入BikaAPI,并适配阅读

+++ ComicX/src/main/java/com/qiuchenly/comicparse/Modules/ComicDetailsActivity/Fragments/ComicBasicInfo/BasicInfo.kt
@@ -8,21 +8,29 @@ import android.view.View
 import com.qiuchenly.comicparse.BaseImp.AppManager
 import com.qiuchenly.comicparse.BaseImp.BaseFragment
 import com.qiuchenly.comicparse.Bean.ComicBookInfo_Recently
+import com.qiuchenly.comicparse.Core.Comic
+import com.qiuchenly.comicparse.Http.BikaApi.ComicDetailObject
 import com.qiuchenly.comicparse.Modules.ComicDetailsActivity.Interface.ComicDetailContract
+import com.qiuchenly.comicparse.Modules.ComicDetailsActivity.ViewModel.ComicInfoViewModel
 import com.qiuchenly.comicparse.Modules.MainActivity.Activity.MainActivityUI
 import com.qiuchenly.comicparse.Modules.MainActivity.Fragments.ComicDashBoard.Recommend.Beans.HotComicStrut
 import com.qiuchenly.comicparse.Modules.ReadingActivity.ReadPage
 import com.qiuchenly.comicparse.R
-import io.realm.Realm
 import kotlinx.android.synthetic.main.fragment_commic_basic_info.*
+import kotlinx.android.synthetic.main.item_comic_infomation.*
 
-class BasicInfo : BaseFragment() {
+class BasicInfo : BaseFragment(), ComicDetailContract.ComicInfo.View {
+    override fun SetBikaInfo(comic: ComicDetailObject?) {
+        tv_comicName.text = comic?.title
+        author.text = comic?.author
+    }
 
     companion object {
         private var mBasicInfo: BasicInfo? = null
         fun getInstance(view: ComicDetailContract.View, comicDetails: HotComicStrut): BasicInfo {
-            if (mBasicInfo == null) mBasicInfo = BasicInfo().apply {
-                mView = view
+            if (mBasicInfo == null) mBasicInfo = BasicInfo()
+            with(mBasicInfo!!) {
+                this.mView = view
                 this.comicDetails = comicDetails
             }
             return mBasicInfo!!
@@ -33,10 +41,14 @@ class BasicInfo : BaseFragment() {
         }
     }
 
+    var mViewModel: ComicInfoViewModel? = ComicInfoViewModel(this)
     var comicDetails: HotComicStrut? = null
     override fun onDestroyView() {
         super.onDestroyView()
+        mViewModel?.cancel()
+        mViewModel = null
         mView = null
+        mBasicInfo = null
     }
 
     private var defaultUrl = ""
@@ -44,19 +56,24 @@ class BasicInfo : BaseFragment() {
         defaultUrl = default
     }
 
-    private val realm = Realm.getDefaultInstance()
+    private val realm = Comic.getRealm()
     private var mView: ComicDetailContract.View? = null
     override fun getLayoutID() = R.layout.fragment_commic_basic_info
 
     @SuppressLint("SetTextI18n")
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        val isBika = comicDetails?.isBika == true
+        if (isBika) {
+            mViewModel?.getComicInfo(comicDetails?.BookLink)
+            return
+        }
+
         test.setOnClickListener {
             val arr = realm.where(HotComicStrut::class.java).findAll()
             realm.beginTransaction()
             arr.deleteAllFromRealm()
             realm.commitTransaction()
-
         }
         addFav.setOnClickListener {
             val book =
@@ -88,9 +105,6 @@ class BasicInfo : BaseFragment() {
                 var obj = realm.where(ComicBookInfo_Recently::class.java)
                         .equalTo("BookName", comicDetails?.BookName)
                         .findFirst()
-                realm.beginTransaction()
-                obj?.deleteFromRealm()
-                realm.commitTransaction()
                 realm.beginTransaction()
                 if (obj == null) {
                     obj = realm.createObject(ComicBookInfo_Recently::class.java, comicDetails?.BookName)
@@ -138,7 +152,5 @@ class BasicInfo : BaseFragment() {
 
     override fun onDestroy() {
         super.onDestroy()
-        if (!realm.isClosed)
-            realm.close()
     }
 }
